Removed nuxt axios module
